Middleware logic fix
If the requested file is inside a folder that is handled by an asset compiler but the extension does not match the compiler's destinationExtension, the middleware will now try to find a counterpart with the same extension inside the assets source folder. If it exists, it will copy the source file to the requested path and serve it.

This is helpful for e.g. handlebars templates or any other assets that should be placed inside an assets folder